Change tensor type to double when loading data

+++ lab/processes/train_model_pytorch/densenet.py
@@ -38,7 +38,7 @@ def load_data_splits(dir_processed: str, batch_size: int, n_workers: int) -> Tup
     data = dict()
     for fname in ["X_train", "X_val", "X_test", "y_train", "y_val", "y_test"]:
         fpath = f"{dir_processed}/{fname}.pt"
-        data[fname] = torch.tensor(torch.load(fpath))
+        data[fname] = torch.tensor(torch.load(fpath)).double()
 
     # Convert tensors to dataloaders
     dataloader_args = dict(batch_size=batch_size, num_workers=n_workers, shuffle=True)
